F OpenNebula/appmarket-simple#6: Add a template viewer
