+++ src/public/react/.babelrc
@@ -3,5 +3,14 @@
     ["@babel/preset-env"],
     ["@babel/preset-react", { "runtime": "automatic" }]
   ],
-  "plugins": []
+  "plugins": [
+    [
+      "@emotion",
+      {
+        "autoLabel": "always",
+        "labelFormat": "[local]",
+        "sourceMap": true
+      }
+    ]
+  ]
 }
