+++ src/public/react/package.json
@@ -39,6 +39,7 @@
     "@babel/eslint-parser": "^7.28.0",
     "@babel/preset-env": "^7.28.3",
     "@babel/preset-react": "^7.27.1",
+    "@emotion/babel-plugin": "^11.13.5",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.6.1",
     "@types/node": "^22.13.1",
     "babel-loader": "^10.0.0",
